+++ packages/typescript-estree/src/ts-estree/ts-nodes.ts
@@ -1,4 +1,4 @@
-import * as ts from 'typescript'; // leave this as * as ts so people using util package don't need syntheticDefaultImports
+import * as ts from 'typescript';
 
 export type TSNode = ts.Node &
   (