+++ packages/typescript-estree/tests/lib/convert.ts
@@ -1,7 +1,7 @@
 // deeplyCopy is private internal
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Converter } from '../../src/convert';
-import ts from 'typescript';
+import * as ts from 'typescript';
 
 describe('convert', () => {
   function convertCode(code: string): ts.SourceFile {